[Refactor] CancelModal 폴더 및 소소한 부분 리팩토링 #471
